add HashExpression method to meta tag record type
