+++ idx/memory/meta_tags.go
@@ -1,10 +1,6 @@
 package memory
 
 import (
-	"hash"
-	"hash/fnv"
-	"strings"
-
 	"github.com/grafana/metrictank/errors"
 	"github.com/grafana/metrictank/expr/tagquery"
 )
@@ -13,18 +9,10 @@ import (
 // slot that's free if two record hashes collide
 var collisionAvoidanceWindow = uint32(1024)
 
-// the function we use to get the hash for hashing the meta records
-// it can be replaced for mocking in tests
-var queryHash func() hash.Hash32
-
-func init() {
-	queryHash = fnv.New32a
-}
+type recordId uint32
 
 // list of meta records keyed by a unique identifier used as ID
 type metaTagRecords map[recordId]tagquery.MetaTagRecord
-
-type recordId uint32
 
 // upsert inserts or updates a meta tag record according to the given specifications
 // it uses the set of tag query expressions as the identity of the record, if a record with the
@@ -37,7 +25,7 @@ type recordId uint32
 // 4) Pointer to the metaTagRecord that has been replaced if an update was performed, otherwise nil
 // 5) Error if an error occurred, otherwise it's nil
 func (m metaTagRecords) upsert(record tagquery.MetaTagRecord) (recordId, *tagquery.MetaTagRecord, recordId, *tagquery.MetaTagRecord, error) {
-	id := m.hashMetaTagRecord(record)
+	id := recordId(record.HashExpressions())
 	var oldRecord *tagquery.MetaTagRecord
 	var oldId recordId
 
@@ -71,21 +59,6 @@ func (m metaTagRecords) upsert(record tagquery.MetaTagRecord) (recordId, *tagque
 	}
 
 	return 0, nil, 0, nil, errors.NewInternal("Could not find a free ID to insert record")
-}
-
-// hashMetaTagRecord generates a hash of all the queries in the record
-func (m *metaTagRecords) hashMetaTagRecord(record tagquery.MetaTagRecord) recordId {
-	builder := strings.Builder{}
-	for _, query := range record.Expressions {
-		query.StringIntoBuilder(&builder)
-
-		// trailing ";" doesn't matter, this is only hash input
-		builder.WriteString(";")
-	}
-
-	h := queryHash()
-	h.Write([]byte(builder.String()))
-	return recordId(h.Sum32())
 }
 
 // index structure keyed by tag -> value -> list of meta record IDs
